herbaria: move 2 forms to Views/
Per #4331:
- Components::HerbariumForm
    -> Views::Controllers::Herbaria::Form
- Components::HerbariumCuratorRequestForm
    -> Views::Controllers::Herbaria::CuratorRequests::Form

ERB callers updated (3). ModalTurboForm dispatch finds Herbarium::Form
via dual lookup (#4328 + #4330). Also updates a HerbariumForm example
in application_form.rb's derive_form_id docstring to point at
Components::NameForm (which is still in components/) since
HerbariumForm has moved.
