+++ app/views/controllers/herbaria/curator_requests/new.html.erb
@@ -7,7 +7,7 @@ add_context_nav(herbaria_curator_request_tabs(herbarium: @herbarium))
 
 <%= :show_herbarium_curator_help.tp %>
 
-<%= render(Components::HerbariumCuratorRequestForm.new(
+<%= render(Views::Controllers::Herbaria::CuratorRequests::Form.new(
       FormObject::HerbariumCuratorRequest.new,
       herbarium: @herbarium,
       back: @back,
