+++ app/views/controllers/herbaria/edit.html.erb
@@ -3,7 +3,7 @@ add_edit_title(@herbarium)
 add_context_nav(herbarium_form_edit_tabs(herbarium: @herbarium))
 %>
 
-<%= render(Components::HerbariumForm.new(
+<%= render(Views::Controllers::Herbaria::Form.new(
       @herbarium,
       user: @user,
       local: true,
